Proc jit-wrapper support
This allows one to use a jit-wrapper interface for procs.

Note that there are some major limitations. If one wishes to use opt_ir one needs to be sure that all required procs are connected and manually communicating with one another via their next methods; having one proc spawn another is not sufficient.

All procs present in the package are ticked when a tick is requested.

The `tick` method is somewhat badly named as it returning success simply means that at least one proc made some progress. It is possible that no procs competed a full tick.

Bug: https://github.com/google/xls/issues/851
Fixes: https://github.com/google/xls/issues/1402
PiperOrigin-RevId: 632262095