Separate aot-compilation and wrapper generation.
The xls_ir_cc_library functioned by basically just calling out to aot_compiler which would compile and write an object file with the compiled version of a dslx function and also write a minimal cc and h file that wrap it.

Move these functions to separate binaries. This will let us make use of the object code dumping in more varied circumstances.

PiperOrigin-RevId: 634001353