Get ready for latest protobuf version.
The protobuffer error collector interfaces changed to accept
string_views instead.
Since these are virtual methods, we need to override
both in the transition period (but not add any 'override' to
it as this depends on the available version we override).

Once we are on latest protobuf (i.e. once we can update grpc),
this will automatically compile with the new version.
After that, we can remove the old AddError()/AddWarning()
methods.

Issues: https://github.com/google/xls/issues/1408
PiperOrigin-RevId: 634099367