Remove all hard-coded pointers from JIT.
The JIT would hard-code pointers to some XLS functions in order to call them during jit execution. This made the code impossible to run outside of the process that created the JIT since the functions may be at different addresses.

To fix this these functions are changed to be passed into the jitted code as a vtable argument in the instance_context argument. This ensures that the jit-ed code can still access all the runtime support it requires.

To avoid embedding Type* pointers into the code all type information is transmitted between the JIT code and the runtime as binary protos.

Bug: https://github.com/google/xls/issues/1403
PiperOrigin-RevId: 635583402