DSLX: enable cross-activation token-ordering semantics
Before this, DSLX did not support any way to express the constraint that "in this proc, operation A in activation 1 must execute before operation B in activation 2". All procs took a single token and returned a single token, and in order to allow pipelining of side-effecting operations, XLS codegen did not consider the next activation's token to depend on the previous activation's token.

In this change, we remove the special token parameter; instead, tokens can be included in the state, and behave like any other value that happens to take up 0 bits (but cannot be optimized away). During execution (and after codegen), operations depending on a given token from the state must wait until after the previous activation resolves that part of the state.

For now, we will continue to accept code in the old style; we will soon:

1. update all of our examples,
2. switch to a mode where the autoformatter will correct code to the new style,
3. block old-style code (with an option for the autoformatter to continue correcting code), and then
4. remove support for the old-style token parameter on proc-next functions.

Fixes #1401 (with some cleanup left over)

PiperOrigin-RevId: 635794009