Enable layering_check and parse_headers where possible.
Not all packages work yet with layering_check, which is
mostly related to insufficient protobuffer dependencies;
It would require to add @com_google_protobuf//:protobuf_nowkt,
but that is deprecated and to be changed soon once we're on
the latest version of protobuffer.

So make these a TODO(google/xls#1411) for now and postpone
until we have updated to the latest grpc and protobuffer.

Issues: https://github.com/google/xls/issues/1411
PiperOrigin-RevId: 635872277