Unify how we prefix the includes that should have project names in path.
There are some external dependencies that have header locations that
have no specific prefix and are somewhat confusing if we used
them as such in XLS:

  * Verible has common/lsp/*.h but ideally we want to clarify that
    these headers come from verible not from some common xls directory.
  * Similar z3. The headers are in the form of src/api/z3_api.h

To include the project prefix, we dealt with this differently for
these projects: In verible, we prepended external/verible (which
then results in header like external/verible/common/lsp/lsp-protocol.h,
which is a valid way to reach the header in bazel).

In z3 we used to prepend it with ../z3, to get header paths in the
form ../z3/src/api/z3_api.h. This adds the project in the header path
but with the slightly awkward relative path.

Unify these two approaches now, by using external/verible and external/z3
as prefix, removing the awkward relative path for z3.

PiperOrigin-RevId: 637964741