Add support for default-driver and load cells in the Yosys synthesis service
Before this, the timing-characterization scripts omitted these parts of the configuration, which could lead to poorly-handled fanout near primary inputs/outputs in the synthesized circuits - producing poor delay characterization for affected operations (including `select`).

This was the root cause for google/xls#1461, which will be fixed once we regenerate the delay model.

PiperOrigin-RevId: 641218886