Make sure [#quickcheck]'s are also filtered by --test_filter
The test filter filtered just the unit tests, but then proceeded
to ran all quickchecks. If there is a filter in place, names of
quickchecks should also be subject to that filtering.

This makes it possible to filter out a single test to run, which
would open the possibility to have a 'Play button' next to
tests in editors.

PiperOrigin-RevId: 642066366