Codegen verilog using verilog types specified by the ir-interface.
This makes the verilog generated by XLS refer to wires by the verilog type name specified by #[sv_type("<name>")] in the DSLX. The user must be sure to put a dependency on the type definition.

This is done by adding an 'ExternType' to vast which is emitted just as the given extern type name.

Bug: https://github.com/google/xls/issues/1239
PiperOrigin-RevId: 642290611