Make debug message not mess-up user trace_fmt() outputs.
If tests are run with --alsologtostderr (as described in
https://google.github.io/xls/dslx_std/#trace_fmt),
some XLS internal message also shows up for every test,
which is confusing. Make this non-actionable message a
VLOG(), so that all that shows up are actually relevant
messages for the test.

Working towards making it possible to easier use tests and
trace_fmt!() and run tests more easily.

PiperOrigin-RevId: 642514820