[DSLX:TS] Error on ref to unbound parametrics.
Previously we'd get an internal error by running head-first into the
bytecode interpreter and it returning us an unbound error from its execution.

This institutes a look-before-you-leap: we look for freevars from the
parametric expression that are present in the parametric bindings, and check
that those are present in the parametric environment map. If they are not, we
provide a more precise and helpful error message.

Ideally we'd be even more aggressive about reporting parametrics that are
unbound at the point of instantiation, that is what #1495 is for and why
there's a bit of code stubbed out with a TODO in this PR -- there is some
investigation needed for the examples that fail with that enabled, but this
change strictly moves the ball forward by resolving the issue seen in #1473.

Fixes #1473
One step towards #1495