Narrow proc-state elements constrained by signed comparisons.
A relatively common pattern in XLScc code is for loops like

```cpp
for (int i = 0; i < kLimit; ++i) { ... }
```

This leads to a proc where there is a 32-bit state element representing the loop variable `i`. This state element has an update function which is guarded by a 'SLt' which is used by back propagation to discover an interval-set of `[[0, kLimit], [INT_MIN, -1]]` for the state element. Since this interval includes all bits potentially set we were unable to narrow this state element.

With this we use the fact that most update functions for state elements are relatively dense and the limited number of intervals we track in an interval set to do state exploration on the state element, assuming that each segment of the interval set contains a grouping of states. Since we can discover that no path starting from the initial value leads to the `[INT_MIN, -1]` segment of the interval set we can remove it and then narrow the state element to simply the kLimit segment.

PiperOrigin-RevId: 645418633