Narrow proc-state elements that mostly decrement.
Due to the fact we use back-propagation to discover limits for state elements our analysis will fail to discover useful ranges for states which mostly move in a negative direction.

For example if we have a state machine:

```
X_init = 45
X' = X - 1 if X > 4
X' = 45    if X == 0
```

We would discover that the range of `X` is [[4, INT_MAX]]. This is not useful for narrowing. The true range however is [[4, 45]] which might allow narrowing.

To discover this better range we use the interval set segment walking used for negative numbers but split the initial interval into 3 with one segment each for any values above or below the starting value. The walking now starts with the precise starting value only. This causes an additional run of the query engine but is generally worthwhile.

PiperOrigin-RevId: 645423148