FifoConfig: Make comparison operators simple members not friends.
'friend' operators would need to be defined outside,
and in fact clang-tidy has an issue with these as-is:
xls/ir/channel.h:60:15: error: friends can only be classes or functions [clang-diagnostic-error]

The more canoncial way is to just have a single-parameter
operator that compares this with that. So let's do that instead.

PiperOrigin-RevId: 646223646