Don't use `uint`, use `unsigned int` or `uint32_t`
`uint` is not really a defined type in C++, and it just might compile
by accident currently due to some typedef somewhere in the headers.
The canonical type is 'unsigned int', so use that.

Uses of uint should probably be looked at w.r.t. limited precision, and
if uint64_t makes more sense, but that is out of scope for this
simple cleanup.

While file open, fix a std::move() opportunity in status_builder.

PiperOrigin-RevId: 646246729