Fix several issues in riscv_simple.x.
- Endianness: RISC-V ISA specs doc says that the base ISA has a little-endian memory system.
The current riscv-simple.x implementation loads/stores data in big-endian order.
This change changes the endianness to little-endian.

- Endianness test: This change adds tests testing whether load/store instructions correctly
loads/stores data in the little-endian order.

- Generatinng I-format instruction: We cannot determine whether an instruction is an I_ARITH
instruction or an I_LD instruction based on funct3. This change adds a parameter specifying
whether the instruction is an arithmetic instruction or a load instruction.

- Store instruction: RISC-V ISA specs doc says that, for the store instruction, the rs1 reg
has the base address, and the rs2 reg is the source reg. The implementation in riscv_simple
mixed up the two registers. This change fixes this issue.

PiperOrigin-RevId: 646619523