Don't name struct-members `stdout` and `stderr`: they are global macros.
(Struct in question: xls::SubprocessResult).

These names are global values that are, due to C inheritance,
a non-namespace macro. As a macro can expand to anything
arbitrary, this has the potential to create head-scratching
compile errors.

Luckily, the standard library (at least on Linux, with gcc
and clang) has the macro just expand to the name itself,
so it does not create an issue (for us, currently):

```
#define stdout stdout
```

However, whenever the language server (clangd) or clang-tidy sees
these it is confused whether to include stdio.h and eagerly suggests
it. To avoid, we would have to plaster NOLINT at each use.

In the quest to be able to check our code with automatic means without
false positives and negatives, as well as keeping the code readable
without tool-specific comments, let's sidestep that by renaming
the captured content of these streams `stdout_content`, `stderr_content`.

PiperOrigin-RevId: 646629870