Update or-tools and grpc dependency to latest.
They need to be updated in tandem as both rely on a newer version
of protobuf that is incompatible with the old.

The patches in or-tools are not needed anymore as
  * They have fully migrated to absl logging and removed a compatibility
    flag
  * The switching of _not_ using some dependencies works and don't
    have to patched out manually anymore.
    (we set these in our .bazelrc).
    Subsequently: also remove scip and glpk from our dependency
    files as they are not needed anymore.

Affects issue #1408 as the new protobuf version means we don't need
compatibility workarounds anymore (but will be addressd in a
separate change).

PiperOrigin-RevId: 647054928