Use constant-derived values as a heuristic for split points in proc parameter state
Proc state narrowing depends on being able to partition the state space (represented as interval-sets) into distinct segments which are then walked to determine which segments are reachable. This can cause problems if segments do not end up sufficiently broken up. To remedy this we find values that the state element can take which are influenced only by literal values and split the ranges at these points. This is under the assumption that these literals probably represent reset points of some sort and so are likely to be at an extreme value of the state element.

To avoid splitting the state space too much and avoid carrying around huge amounts of memory we limit the analysis to only keep track of the 8 values closest to 0 (in the 2s-complement signed space).

PiperOrigin-RevId: 647339704