Simplify jit/interpreter switch build rules and remove duplicate test.
The emergency switch to an interpreter has been used once, but not really
needed by default, yet we still had a complicated alias-setup both variants.

Keep the functionality in place (it is still quick to change the #define in
switchable_function_jit in case it is needed), but remove the unnecessary
duplication of library and test as it became more of noise
(also, it emitted a loud warning on every test run).

(Removes the interpreter_fallback_test, but that is really just a mostly
verbatim copy of the non-fallback test anyway).

PiperOrigin-RevId: 647728397