Use max_frequency_hz as delay in GrpcSynthesizer when valid.
Both Yosys and DC gRPC services report a valid max_frequency_hz, and we will now use that in practice with both. The Yosys service also returns a valid slack_ps vs. target frequency, and the API theoretically allows support for only that approach, so we will keep supporting it.

PiperOrigin-RevId: 648437667