Extend conditional specialization pass to forward through OneHotSelects & logical ops too
We already supported forwarding cases through case-at-a-time selects where the context guaranteed which case would be selected.

We now extend this logic to also forward inputs through OneHotSelects & ANDs/ORs/XORs where the context guarantees that one input passes identically through the operation. Applied to AND/OR/XOR, this should give us a chance to see through various forms of operations that amount to selects or gates in context.

PiperOrigin-RevId: 648509568