Fix identity-checking in conditional specialization pass
The pass was not correctly tracking which value was the identity for a given bitwise operation. This could lead to incorrect results if attempting to "see through" a sequence of mixed operations.

Adds a test to verify that we don't regress in the future.

Fixes recent fuzzer errors, including `0031623b` & `f44c236e` (from `opt_function_narrow_nosim`)

PiperOrigin-RevId: 649147052