Implement the block stitching pass, the core component of multi-proc codegen.
The block stitching pass creates a top-level container block that instantiates all the other blocks. Send/receive channels between blocks are stitched together according to the FIFO configuration of the channel- if depth=0, the channel's ports are directly connected, otherwise a FIFO is instantiated and the push and pop sides are connected to the appropriate block's ports. Send-only and receive-only channels are punched out of the top-level container block.

Currently, only streaming channels are supported.

Work towards #950.

PiperOrigin-RevId: 653757118