dependency_support: Remove net_zstd
net_zstd dependency is not referenced anywhere in XLS.
It was introduced as a transitive dependency of fuzztest.
In fuzztest workspace net_zstd is also a transitive dependency of riegel.
Riegel version used in fuzztest bazel workspace is WORKSPACE-based so transitive dependencies must be defined in workspaces that use that version of riegel.
XLS however uses newer, bzlmod-based version of riegel which does not require that.
That means there is no need for @net_zstd dependency in XLS workspace and it can be removed.

Signed-off-by: Pawel Czarnecki <pczarnecki@antmicro.com>