Add an XLS stat marker to yosys tcl script; parse area stats
The stat marker helps the stat parsing regexes identify the region
of yosys output containing the stats of interest.

As of now, using the yosys synthesis service will result in two sets
of stats. The first set is outputted by the synth step, which does not
have the area stats. The second set is outputted by a manual `yosys stat`
call, which only outputs that stats of the top module. We use this marker
to indicate that start of the second set of the stats, which contains the
cell histogram and area stats of the top module.

PiperOrigin-RevId: 654942618