Use static regexes for parsing yosys outputs
Since the regexes are known at compilation time and are constants
during runtime, we opt to store the regex objects as `constexpr static`
objects.

Since RE2 is not safe for static use, we opt to use LazyRE2, which is
safe for static use. As a side note, LazyRE2 objects will be lazily
constructed during runtime.

PiperOrigin-RevId: 655346007