Include some missing headers and dependencies.
Usually, we would not include absl/strings/string_view.h, however
in this case we actually do use a function proivded from that header
(and _not_ the absl::string_view type).

In log-lines we're uing an internal definition ABSL_RAW_CHECK
in absl for which we can't include the header (as the corresponding
library is not public) but at least document that.

While at it, fix a couple of missing includes and dependencies that
have crept in recently.

PiperOrigin-RevId: 657657145