Include poll header that clang-tidy is happy with.
The <poll.h> only redirects to <sys/poll.h> on Linux, which
makes clang-tidy unhappy as it actually does not defined symbols
_directly_. So include sys/poll.h instead (and let's assume that this
work on other UNIXes as well).

WEXITSTATUS and WIFEXITED are defined by stdlib.h, but it is not
seen by clang-tidy when we include the 'modern' header <cstdlib>.
So include the traditional one instead.

PiperOrigin-RevId: 658197772