Update DFE for multi-proc codegen.
This changes DFE to consider all send_only and receive_only channels as side-effecting regardless of if they are reachable by the top proc. The idea is that some programs have multiple procs that don't communicate with each other but are observable via external (i.e. send_only or receive_only) channels.

This is useful for multi-proc codegen (google/xls#950) where we codegen more than the top proc.

PiperOrigin-RevId: 658640856