Add helper for generating a main.cc runner for an AOT entrypoint.
This just blindly generates a test main.cc file which invokes the compiled code and compares its result to an expected result, optionally writing the result. This cc file intentionally has no runtime dependencies (not even libc) to enable it to be used for debugging llvm issues using LLI and other tools.

PiperOrigin-RevId: 660104292