During type deduction of tuples, when a "rest of tuple" (`..`) operator is seen, skip ahead the correct number of entries in the tuple values array.
This change only updates the type system via type deduction.  It does not address bytecode generation or IR conversion for the new syntax.

As of this change, using "rest of tuple" will still fail in the interpreter if it needs to skip multiple tuple values. If it matches exactly one value, it will succeed, and acts as wildcard (`_`).

https://github.com/google/xls/issues/1459

PiperOrigin-RevId: 660818563