Update `FunctionConverter` to generate IR when there is a "rest of tuple" operator (`..`) in a `let` statement.
Implementation details: the processing here is opposite from `Deduce` and
`BytecodeEmitter`; those two *recursively* process a whole tuple level at a
time, so we can adjust the index into the tuple values when we see a "rest of
tuple" `NameDef`.

For `FunctionConverter`, it walks *iteratively* via a callback (`DoPreorder`)
through the `NameDef`s and assigns the tuple index the same value as the
`NameDef` index. Now, to correspond to the recursion of the other methods,
a list of "deltas" (one per level of the tuple) is maintained. After seeing a
"rest of tuple" `NameDef` at a particular level, the delta is added to the
index in order to adjust the index of each subsequent element at that level.

This does not address `match` yet.

https://github.com/google/xls/issues/1459

PiperOrigin-RevId: 662549474