[XLS] Reorder selects to encode some functions of the selector
Any time we compute `sel(f(x), cases=[...])`, it's equivalent to `sel(x)` for some permutation of the cases. When it's locally visible that the selector is a unary function of some other node, we now take advantage of this & encode the function into the select's cases, moving the computation to compile-time.

PiperOrigin-RevId: 665403334