Actually unroll unroll_for! loops during type checking.
Previously, this was never being done. The practical use of channel arrays currently requires explicit unrolling.

https://github.com/google/xls/issues/704

Fixes #704

Because a loop needs different unrolled renditions per parametric env, we store those with the corresponding env in TypeInfo, rather than trying to produce an edited AST, as old comments incorrectly implied was happening. This means a visitor that wants to use unrolled loops does in fact need trivial handling for UnrollFor, which retrieves the desired unrolled rendition, and then recurses into it.

PiperOrigin-RevId: 665884726