Make Ops less generated-code-y
These were generated using python previously but since removing that the checked in code has a lot of code-generator-isms. Remove them and make it generally more hand-written-y and easier to modify manually.

Most notably this removes the need to forward declare every node type and puts all the attributes of an op together in a single macro.

PiperOrigin-RevId: 665989424