Add helper for generating LLVM Artifacts
We have at various points found several bugs in the LLVM code-generation/optimization pipeline. These are typically revealed by some XLS code which generates odd LLVM ir when jitted/aot'd. Previously it has always been a laborious and manual process to collect everything together in order to do validation that the issue is in LLVM and to collect the artifacts needed to submit an LLVM bug report.

To simplify this process a new xls/jit:dump_llvm_artifacts binary is now available which can be given an XLS function IR file, an input set and an expected output and dump a collection of LLVM ir files which implement a basic tester. These files include a 'linked.ll' and 'linked.opt.ll' file which is a single llvm ir file with a 'main' method that exits '0' if the XLS code results in the expected outcome and '1' otherwise. This can easily be used with --bisect_passes_limit in llvm-opt to determine where a miscompile is ocuring.

PiperOrigin-RevId: 666472428