Fix issue with Proc-state-narrowing.
In cases where two 'next_value' ops have the exact same value parameter but one of them is provably never enabled a bug in segment-walking can occur. What happens is segment walking tries to avoid merging intervals when it does not have to so if a value has already been merged into the interval it will continue to the next 'next_value' instruction early. Unfortunately the value is marked as handled before actually checking whether we will merge it into the intervals. If the 'next_value' nodes are iterated such that the un-activatable one goes first this can lead to incorrect calculation of possible values.

PiperOrigin-RevId: 666895571