Rename xls/dev_utils to xls/dev_tools. This is in preparation of moving developer-only tools from xls/tools to xls/dev_tools. This rename makes the names of the directories more consistent.
Also, in copybara, remove line which moves dev_tools (formerly dev_utils) up to the top level of the repo. Previously this was done with convenience in mind (?), but now the directory includes many tools and libraries which are depended on by other directories so moving the directory is cumbersome and doesn't make sense.

PiperOrigin-RevId: 667769405