Bit provenance Tracking
This analysis allows one to track where individual bits of a node come from. In most cases a node will be considered the source of all of its bits but in the case of concats, and slices and so on the source might be some other node. This can be leveraged to allow narrowing even in cases where the actual value is difficult to determine.

PiperOrigin-RevId: 668040566