[XLS] Add `FillPattern` to node_util, and extend `GatherBits` to take a mask
`FillPattern` is a utility function that takes a pattern of known and unknown bits and a node, and returns a new expression that matches the given pattern where known and fills in with bits from the node otherwise.

`GatherBits(node, mask)` is the inverse of `FillPattern`, in the sense that `GatherBits(FillPattern(pattern, node), ternary_ops::ToKnownBits(pattern))` will always have a result equivalent to `node`.

Applies `FillPattern` to simplify the narrowing-pass implementation for literal array narrowing.

PiperOrigin-RevId: 668093911