[XLS] Add optimization for dynamic slices with select-of-literal starts
If a dynamic bit slice or bit-slice update has a select of literals for its start position, hoisting the operation above the select lets us convert it to a select of static operations.

While we're here, add a simplification to identify known out-of-bounds dynamic bit slices and replace them with zeros.

PiperOrigin-RevId: 668917814