[XLS] Avoid invoking the SDC scheduler for single-stage pipelines
When all the nodes are specifically going into a single stage, there's no need to start up an expensive scheduler. Even if the solver will find the only viable solution quickly, encoding the problem has a nonzero cost!

Also add support for MinDelay to the ASAP scheduler, since that will be used in these scenarios.

PiperOrigin-RevId: 671113884