Fix hash collisions among data points when using timing_characterization_client; fix #1525
Currently, there is a hash collision of two kArrayIndex data points when running
timing_characterization_client. This caused by the fact that the timing client
relies on hashing the op to map a sample to its result, and the hash function did
not take into account the element count of every operand. So, if there are two
data points that only differ in element counts, the data point of only one of the
two samples will be recorded in the output.

This change fixes the problem by adding element counts part of the hash.

PiperOrigin-RevId: 671152475