Use OpNameRepresentation proto to map cpp enum names to op names
Recent changes remove the map mapping cpp enum names to the op names. As a result,
we're relying on a heuristic in which the translation being a transition of camel
cases to snake cases. However, this heuristic fails for the signed/unsigned ops,
e.g., 'kUMul' should be mapped to 'umul' instead of 'u_mul'. This heuristic breaks
delay model data point generation.

This change uses the map provided by the proto for translation.

PiperOrigin-RevId: 671827523