Fix incorrect narrowing of multiply.
In some situations it is possible to narrow a multiply all the way to a constant zero (if the number of trailing bits on both operands exceeds the number of bits in the result). In this case we incorrectly tried to generate a multiply which results in a negative bit-width result causing a CHECK failure.

PiperOrigin-RevId: 671927445