Improve abstract-evaluator handling of select-ops
The abstract evaluator is constructed to use boolean logic to implement various operations. This includes implementing an implementation of (a 1 bit-wide)'if' that looks like `(or (and selector consequent) (and (not selector) alternate)`. This uses the law of the excluded middle to ensure that either the consequent or alternate are selected. Unfortunately the Ternary abstract evaluator violates the excluded middle since it does not keep track of the fact that the two uses of `selector` are linked. This means the ternary evaluator is unable to see some quite simple facts. This was most prominent on priority-select but could occur with other selects as well (though not with the ternary evaluator).

This adds a 'If' primitive operation to the abstract-evaluator which implementers can use to ensure the excluded middle is respected to the largest extent possible.

PiperOrigin-RevId: 671938752