Simplify & fix the AbstractEvaluator implementation of PrioritySelect
Before this, we were trying to do too much in terms of optimizations & ended up with a segfault if `cases` was empty. We can just convert the priority select into a set of nested ternaries in the opposite order, and cut off the default if we know the selector can't be zero.

PiperOrigin-RevId: 673432452