Speed up z3 models.
We were using the abstract evaluator to generate z3 models for selects and bit-slice-updates (which internally are implemented as selects in the abstract evaluator). This is rather inefficient since it creates a lot of shattering of bitvectors and a huge amount of repeated expressions for each bit in the select. Instead we can just implement these directly using z3 'ite' (if-then-else) operations. This substantially speeds up sat-checking of these models. Overall this more than doubles the speed of some xlscc and codegen targets.

PiperOrigin-RevId: 673554915