Implement area model
Additional notes,
- The op_models and samples of the area model are the same as those in the delay model.
However, the op_models and the samples of the area model should be modified as it is
apparent that the parameterization and asymptotic growth of the area model of every op
might be very different from the counterparts in the delay model.
- The AreaEstimator class has a function to get one-bit register area. This area is
determined and populate when the derived classes are generated. There is a heuristic
that, if kIdentity is a fixed estimator, the AreaEstimator is a unit area model, and
one-bit register area will be the same as the constant returned by kIdentity. If
kIdentity is not a fixed estimator, then it's must be either RegressionEstimator
or AreaRegressionEstimator. The codegen picks up one of kIdentity data points
and derives the one-bit register area from the sequential area of that data point.

PiperOrigin-RevId: 674050718