[DSLX:TS] Fix mismatch errors for concat to be TypeInferenceErrors.
The fundamental problem is that TypeMismatch errors are semantically
representing "we expected the type to be X but it was Y".

When you try to use that for operands that *are matched, but somehow not
right*, the internal formatting machinery is upset because it's looking
for a discrepancy that should be there as a precondition.

This switches the reporting to be TypeInferenceError, which is the more
correct error in this scenario, and also fixes a likely problem due to
not using `GetBitsLike` adding a test for concatenating with a properly
unsigned `xN`.