Fix incorrect evaluation of spawn args in bytecode interpreter.
The motivating use case is described in https://github.com/google/xls/issues/1502

The modeling of channels as "constexpr" InterpValues through which you can push data at runtime is brittle; these InterpValues should really become channel IDs, with the interpreter then instantiating channels at runtime.

Before this change, the problem was compounded by the fact that the interpreter relied on constexpr evaluation for both `ChannelDecl` nodes and spawn args. It expected spawn arg exprs to constexpr-evaluate to the same "channel InterpValue" as the appropriate `ChannelDecl`. But when you have multiple test procs in a file, one proc network can taint the constexpr evaluation of another's spawn args.

The solution is to pass spawn args on the stack, as if `spawn` were a function, so that we only depend on constexpr evaluation of the `ChannelDecl` and we don't have multiple constexpr evaluations of what should be the same channel in the same proc network.

Longer term, proc-scoped channels, and the removal of runtime channel data structures from constexpr evaluation, will further clean this up.

Fixes #1502

PiperOrigin-RevId: 675316200