Simplify sample_runner by keeping commands simple functions.
Instead of a variant that keeps track of either an executable path
or function, always have commands simply be functions.
For calling exectubable, wrap that in a function.

While at it, group the set of constants that point to the
binaries in a struct for improved readability.

Overall, no functional change, just readability improvements.

PiperOrigin-RevId: 675752356