When instantiating an externally-defined struct, peek to see if the user is trying to parameterize its instantiation. Previously, it was only looking for the open brace after an external name, and if it wasn't found, it assumed it was an external function call.
Now, if the peeked token is an open angle bracket (`<`), it tries to do an external parametric struct instantiation. This may still fail (e.g., if it's an external enum comparison, or an array of external parametric structs), in which case it rolls back and continues as before.

Partially addresses https://github.com/google/xls/issues/1030

PiperOrigin-RevId: 676110972