[XLS] Optimize DataflowVisitor by avoiding redundant queries for array indices
When addressing ArrayIndex and ArrayUpdate operations, we used to query the QueryEngine for the same index nodes repeatedly for each possibly-addressed value. We now query for any known index nodes' values at the start, then use them in each case.

While we're at it, we also clean up the code a bit to remove additional redundancies.

PiperOrigin-RevId: 676160297