fuzzer: make run.sh directory independent.
The location of a run might move from the run_dir to the crasher
directory, so the directory to change to should not be hard-coded
but relative to the script.

While at it: pass any additional parameters to the runner, so that
the user could, e.g. choose to invoke with -v=1.

PiperOrigin-RevId: 677876734