Fixed an invalid memory access.
The merge https://github.com/google/xls/pull/1594 created an invalid memory access.
This commit fixes it.

The invalid memory access is created by the following statement:
dslx::FileTable& file_table = globals.dslx->import_data.file_table();
This code accesses the field dslx of the global struct "globals". Such field is a unique_ptr that has not been allocated yet.

The fix brings the creation of the independent object "import_data" before the statement above.
Then, it replaces the above statement to use the import_data so created.
Please notice that both before this fix and after it, the import_data object is transferred to the Constructor of the field "globals.dslx", which is invoked after the target statement.
Therefore, there is no harm in bringing the construction of "import_data" earlier.

fixes google/xls#1623

PiperOrigin-RevId: 678212464