Fix bugs with type checking unroll_for! expressions.
The root cause of the specific case flagged in #1614 is that we were not uniformly applying the type-checking rules that should apply to all for loops. This change factors out the common type checking logic and applies it to both, then DeduceUnrollFor applies extra constraints.

This change also improves some existing error messages to stop talking about the (index, acc) tuple as a unit.

https://github.com/google/xls/issues/1614

Fixes #1614

PiperOrigin-RevId: 678383527