Speed up `IntervalSet::Complement`.
The previous implementation modeled `Complement` as the intersection of the complements of the comprising intervals. However, since we already invoke `Intervals()` in this function, it only applies when the `IntervalSet` is already normalized... at which point, we can be more efficient by scanning linearly through the sorted list of non-abutting intervals & recording the unused values.

PiperOrigin-RevId: 678405288