Tell clang-tidy explicitly to consider code c++20.
Even though the -std=c++20 makes it into the
compilation DB, clang-tidy does not pick it up.

So be explicit and add it to the --extra-args on
invocation.

Thie fixes all reported [clang-diagnostic-error] that
were due to clang-tidy not recognizing c++20 constructs
and also some misc-include-cleaner warnings in which it
previously flagged some includes because it was tripping
over some other problem in the file in question.

<    21 [misc-include-cleaner]
<    20 [clang-diagnostic-error]
---
>    15 [misc-include-cleaner]

PiperOrigin-RevId: 678427605