Allow channel metadata to be referenced from multiple blocks.
In multi-proc codegen, the two sides of a channel have ports in potentially different blocks, but the metadata only reflected one of them (based on execution order). This change makes it so that channel metadata can reference the port names for any block it exists in. As is usual with signatures, this has a significant blast radius.

Progress towards google/xls#1323.

PiperOrigin-RevId: 679743038