[XLS] Add a pass to convert nodes to lookup tables where it's beneficial
This pass is currently only enabled for selectors of `select` operations, where we can guarantee (local) PPA improvement, but it could be extended to other nodes in the future.

For this initial version, we add a new dataflow-dominator analysis that can locate nodes D for any given X such that X's output is entirely determined by D's output. Given a select, we find the earliest dataflow-dominator for its selector that will not produce a wider select; we then evaluate the selector with all possible dominator values, encoding the result into a new select that embeds an equivalent lookup table.

PiperOrigin-RevId: 680749998