Fix infinite loop in type checking with inverted parametric name order.
Context: https://github.com/google/xls/issues/1615

The main scenario we are dealing with here is:

```
struct Foo<X, Y, Z = {X}> { ... }

fn bar<X, Y>() -> Foo<Y, X> {
  // Reverse-ordered parametric names in the instantiation vs. the struct def.
  Foo<Y, X> {...}
}
```

The defaulted `Z` is not needed to hit the bug, but the bug is fallout from previously fixing that, and we want to keep it fixed.

In this scenario, it's OK to evaluate `Z` with the parametric env that has `X` and `Y` from the immediate context in `ConcretizeStructAnnotation`, but we cannot evaluate the externally provided `Y` and `X` with that context, as we were doing, or we end up believing they are circular. So this change limits evaluation to the implicit dims.

The other problem with the scenario is that the nominal struct dim vs. return type check in `TypecheckFunction` would fail, because `Resolve` was using `AddNominalTypeDims` to erroneously add (to the function return type) the parametric env from one "frame" down rather than using it for evaluation. This change breaks out and fixes the resolution step for nominal dims.

Fixes #1615

PiperOrigin-RevId: 681441282